requirements.txt
paho-mqtt
python-dotenv
requests
PyYAML